Use read-only and read/write databases.
Originally the path to a single read/write database was set with the "database" setting. If a game, cheat, or code line was created, modified, or deleted, the entire database was rewritten.

Now there can be a read-only database in addition to a read/write one, specified by the "databaseReadOnly" and "databaseReadWrite" settings, respectively. If a game, cheat, or code line is created, modified, or deleted, it is written to the read/write database rather than rewriting the whole database. The read-only database won't be written to by CheatDevice.

Loading works, still need to implement saving only modified games/cheats to the read/write database.