Add support for value-mapped code type.
These are cheats that corespond to a set of values specified by a map. For example, many games define item inventories as an array of slots, where each slot contains a char to represent a specific item. With this code type you could define all possible values in a map, then reference the map by name in place of the code's value.

This commit implements cheat database parsing for maps and value-mapped code lines, and menu logic to select values.