engine: Fix multiple hook (enable code) support.
Although the pointer to the hook list entry (stored at $t3) is
being updated at the end of the loop (as expected), the pointer
is re-initialized at the next iteration by loading the start
address of the hook list (hooklist) in $t3. Place the initialization
of the pointer outside the loop (as done in ps2rd / ps2rd-gui).