src: Fix false positives at getToken() function.
Any comment line with length of 17 chars and a space as the 9th
char (e.g. //ABCDEF GHIJKLMN) is being mistakenly identified as
a cheat title line (TOKEN_CHEAT), due to getToken() returning
TOKEN_CHEAT instead of continuing with the next (and final)
check (testing if the line starts with "//" or "#", i.e. testing
if the line is a comment).

Also, fix an issue with code lines located below comment lines, as
the former were being misdetected as TOKEN_CHEAT (cheat title)
instead of being detected as TOKEN_CODE (code line).